# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Atlas Platform API
  version: 0.2.0
  description: Platform operations — orgs, onboarding, webhooks, events, reports, exports, and cron jobs
  contact:
    name: AtlasPayments Team
    url: https://github.com/shivemind/atlaspayments
servers:
  - url: http://localhost:3000
    description: Local development server. Override via Postman {{baseUrl}} environment variable.
...
paths:
...
  /api/internal/onboarding/{id}/bank_accounts:
    get:
      summary: List bank accounts
      description: Returns a paginated list of bank accounts for the onboarding profile.
      x-platform:
        kind: resource
        entity: BankAccount
        affectsLedger: false
      operationId: listOnboardingBankAccounts
      tags:
        - Onboarding
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated list of bank accounts.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/OnboardingBankAccount"
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Onboarding profile not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Add bank account
      description: |
        Adds a bank account to the onboarding profile. If marked as primary, any existing primary account is demoted.
      x-platform:
        kind: resource
        entity: BankAccount
        emitsEvents:
          - onboarding.bank_account.created
        affectsLedger: false
      operationId: createOnboardingBankAccount
      tags:
        - Onboarding
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - bank_name
                - account_holder_name
                - routing_number
                - account_number_last4
              properties:
                bank_name:
                  type: string
                account_holder_name:
                  type: string
                routing_number:
                  type: string
                account_number_last4:
                  type: string
                  minLength: 4
                  maxLength: 4
                account_type:
                  type: string
                  enum:
                    - checking
                    - savings
                  default: checking
                is_primary:
                  type: boolean
                  default: false
            examples:
              basic:
                value:
                  bank_name: First National Bank
                  account_holder_name: Acme Corp
                  routing_number: "021000021"
                  account_number_last4: "6789"
                  account_type: checking
                  is_primary: true
      responses:
        "201":
          description: Bank account created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/OnboardingBankAccount"
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Onboarding profile not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/internal/exports:
    get:
      summary: List all exports (admin)
      description: Returns a paginated list of all exports across merchants.
      x-platform:
        kind: resource
        entity: Export
        affectsLedger: false
      operationId: adminListExports
      tags:
        - Exports
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - in: query
          name: merchant_id
          schema:
            type: string
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum:
              - queued
              - processing
              - completed
              - failed
        - in: query
          name: format
          schema:
            type: string
            enum:
              - csv
              - json
              - pdf
      responses:
        "200":
          description: Paginated list of exports.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Export"
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/internal/orgs/{orgId}/roles:
    get:
      summary: List org roles
      description: Returns a paginated list of roles for an organization.
      x-platform:
        kind: resource
        entity: OrgRole
        affectsLedger: false
      operationId: listOrgRoles
      tags:
        - Orgs
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - *a3
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated list of org roles.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401": *a1
        "403": *a2
    post:
      summary: Create org role
      description: Creates a new role for an organization.
      x-platform:
        kind: resource
        entity: OrgRole
        affectsLedger: false
      operationId: createOrgRole
      tags:
        - Orgs
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - *a3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                permissions:
                  type: array
                  items:
                    type: string
      responses:
        "201":
          description: Org role created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
        "400": *a4
        "401": *a1
        "404": *a5
  /api/internal/orgs/{orgId}/roles/{id}:
    get:
      summary: Get org role
      description: Returns a single org role by ID.
      x-platform:
        kind: resource
        entity: OrgRole
        affectsLedger: false
      operationId: getOrgRole
      tags:
        - Orgs
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - *a3
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Org role details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
        "401": *a1
        "404": *a5
    patch:
      summary: Update org role
      description: Updates an org role.
      x-platform:
        kind: resource
        entity: OrgRole
        affectsLedger: false
      operationId: updateOrgRole
      tags:
        - Orgs
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - *a3
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                permissions:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Org role updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
        "400": *a4
        "401": *a1
        "404": *a5
    delete:
      summary: Delete org role
      description: Deletes an org role by ID.
      x-platform:
        kind: resource
        entity: OrgRole
        affectsLedger: false
      operationId: deleteOrgRole
      tags:
        - Orgs
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - *a3
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Org role deleted.
          content:
            application/json:
              schema:
                type: object
                required:
                  - deleted
                  - id
                properties:
                  deleted:
                    type: boolean
                  id:
                    type: string
        "401": *a1
        "404": *a5
...
components:
  securitySchemes:
    MerchantKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for merchant-scoped operations. Pass as `Authorization: Bearer {{merchantApiKey}}`.
    PlatformAdminKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for platform admin operations. Pass as `Authorization: Bearer {{platformAdminKey}}`.
  schemas:
    ErrorResponse:
      description: Standard error response envelope returned for 4xx/5xx responses.
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code.
            message:
              type: string
              description: Human-readable error description.
    PaginationMeta:
      description: Pagination metadata included in list responses.
      type: object
      required:
        - page
        - pageSize
        - total
        - hasMore
      properties:
        page:
          type: integer
          description: Current page number (1-indexed).
        pageSize:
          type: integer
          description: Number of results per page.
        total:
          type: integer
          description: Total number of matching records.
        hasMore:
          type: boolean
          description: Whether additional pages exist.
...
    OnboardingBankAccount:
      description: Bank account attached to an onboarding profile.
      type: object
      properties:
        id:
          type: string
        onboarding_profile_id:
          type: string
        bank_name:
          type: string
        account_holder_name:
          type: string
        routing_number:
          type: string
        account_number_last4:
          type: string
        account_type:
          type: string
          enum:
            - checking
            - savings
        is_verified:
          type: boolean
        is_primary:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Export:
      description: A data export requested by a merchant.
      type: object
      properties:
        id:
          type: string
        merchant_id:
          type: string
        type:
          type: string
        format:
          type: string
          enum:
            - csv
            - json
            - pdf
        status:
          type: string
          enum:
            - queued
            - processing
            - completed
            - failed
        parameters:
          type: object
          nullable: true
        result_url:
          type: string
          nullable: true
        error_message:
          type: string
          nullable: true
        row_count:
          type: integer
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  parameters:
...
    PageParam:
      in: query
      name: page
      description: Page number (1-indexed).
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      in: query
      name: pageSize
      description: Number of results per page (max 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    PathId:
      in: path
      name: id
      required: true
      description: Resource identifier.
      schema:
        type: string
...